Amended user creation
The user had no /home folder on build so x2go fails when trying to chroot into it. Added -m to the adduser command to create this on the next build. 

Also added a desktop note to show users how to give the x2go user sudo access for ease of use.